Add cache variable GEANT4_INSTALL_DATADIR
This variable provides a checking and install path for the
Geant4 data libraries. It is set up along the same lines as those
in the standard CMake GNUInstallDirs module so that there are
relative and absolute paths available.

It defaults to DATAROOTDIR/Geant4Data. This is just a placeholder at
the moment.