Add md5sum to dataset entries
MD5 can be used CMake 2.6.4 and higher, so add hash of file to be
downloaded as the last entry in the tuple for each dataset.

Dataset tuples have be reverted to the existing slash separated format
so it'll be easy to patch changes into current Geant4. However, may
need to change this if we implemented checking for datasets via marker
files.