Refactored data install for CMake 2.6.4
If data is installed, each dataset is installed via a front end function
which dispatches to CMake's ExternalProject if CMake is of sufficient
version, or to an internal _geant4_dataproject function.

Each function downloads, verifies and unpacks the data into the same
structure in the binary directory. Installation is then straightforward.

Suitable defaults for the data installation, checking of existing
installs and setting of variables for use in configuration scripts
needs to be added.