Add full support for downloading from CERN repo.
Use http CERN repo to obtain data with new MD5 sums in dataset tuples.
Remove logging from CMake < 2.8.2 use of file(DOWNLOAD ...) due to bug
in using this with both TIMEOUT and LOG arguments. Use pure TIMEOUT as
this is more relevant to users with slow connections, and we do want a
timout to prevent locking.

A default for the dataset install path is also set as part of the
global data so its changes are easy to make without having to dig
down into the later User Interface section.