Add outline logic block for reuse of datasets
As datasets are large, we want to use any sets that are already
installed in the final install location for datasets. Implement
a block checking the existence of datasets using the following
logic:

- If the dataset already exists, reuse it.
- If the dataset does not exist, and GEANT4_INSTALL_DATA is ON,
  download and install the dataset.
- Otherwise, warn that the dataset is missing and should be
  installed manually.

Only message at the moment for testing, adding the appropriate
tasks later when the dataset API is finished.