Begin implementation of G4data CMake API
Add CMake function for defining datasets using same variable set
as current "tuple" style. Dataset name is added to global property
"GEANT4_DATASETS", with each dataset property set as a global
property with name DATASETNAME_DATASETPROPERTYNAME.

Implement further functions to check if dataset exists and to
return (via global scope setting) values of properties for a
given dataset.

Wrap logic of checking for, reusing if found, and downloading
otherwise (if requested) into a single function. This function
only reports what will happen at the moment.

Note that this is all still a bit of mix of the old style and
new style APIs....