Provide build/install dir properties for datasets
Add new geant4_dataset_set_property function for setting properties
of datasets after dataset has been declared. Property names are
not limited as yet, though should be to ensure clarity and
accuracy.

In dispatch functions for download/install and reuse of datasets,
use the new function to set properties for each dataset which point
to the location of the dataset for use with a build and
final install of Geant4. Name these as:

- BUILD_DIR: the absolute directory path to the dataset when running
  from the Geant4 build directory.
- INSTALL_DIR: the final (absolute) path to the dataset when installed.

If a dataset is reused or configured in absentia, these paths are
identical.

Update geant4_data_create_exports macro to a function. Use this
to set the appropriate @VAR@s in parent scope for each dataset,
and generate both build and install tree export files.