Hacky implementation of exporting shell code block
As number of datasets may evolve with time, we don't want to
hardcode their environment variables into the shell scripts that
export them. This means expanding a @VAR@ whose value is a sequence
of lines, each being an export/setenv for the dataset envvar. This
can be generated on the fly in CMake.

This commit is a first attempt at doing this for the Bourne shell
only. Add a variable to the geant4data.sh.in template that will
expand to the code block. Use the new dataset API to create each
export command from the known datasets. Generation of the final
file proceeds as before.